skip empty fields
